requirements.txt
GeoIP
pytelegrambotapi
requests
PySocks
pillow
